+++ packages/web-components/src/components/global-header/components/global-header/src/components/HybridIpaasHeader/HybridIpaasHeader.ts
@@ -43,7 +43,7 @@ export class HybridIpaasHeader extends LitElement {
   @property({ type: String }) productKey = '';
   @property({ type: Object }) fetchHeaders = {};
   @property({ type: Boolean }) sidekickEnabled = false;
-  @property({ type: Boolean }) SolisEnabled = false;
+  @property({ type: Boolean }) solisEnabled = false;
   @property({ type: String }) basePath = '';
   @property({ type: String }) displayName = '';
   @property({ type: String }) userEmail = '';
@@ -163,7 +163,7 @@ export class HybridIpaasHeader extends LitElement {
   }
   private initSolisOptions() {
     return {
-      isEnabled: this.SolisEnabled,
+      isEnabled: true,
       scriptUrl:
         'https://cdn.dev.saas.ibm.com/solis_ui/v1/switcher/solis-switcher.es.js',
       is_prod: false,
@@ -272,7 +272,7 @@ export class HybridIpaasHeader extends LitElement {
     if (this.sidekickEnabled) {
       updatedOptions.sidekickConfig = this.initSidekickOptions();
     }
-    if (this.SolisEnabled) {
+    if (this.solisEnabled) {
       updatedOptions.solisConfig = this.initSolisOptions();
     }
 
@@ -292,7 +292,6 @@ export class HybridIpaasHeader extends LitElement {
     ) {
       updatedOptions.globalActionConfigs = [...this.capabilityGlobalActions];
     }
-
     return updatedOptions;
   }